add component page layout and button example
