+++ site/docs/components/button/examples.mdx
@@ -1,0 +1,19 @@
+---
+title: Button
+layout: DetailComponent
+sidebar:
+  label: Examples
+data:
+  descriptionRef:
+    $ref: ./#/data/description
+---
+
+### CTA (call to action)
+
+Use the CTA button for high priority actions.
+
+#### Guidance
+
+- Choose a CTA button when an action is more urgent than other visible buttons.
+- Use it for important or irreversible actions, such as agreeing to terms and conditions or deleting a profile.
+- Only display one CTA button per screen to emphasize the next action the user should take.
