+++ site/docs/components/button/index.mdx
@@ -1,0 +1,6 @@
+---
+title: Button
+layout: DetailComponent
+data:
+  description: Button is an interactive component that allows the user to execute an action. There are three Button variants—CTA (call to action), primary and secondary—that you can display with or without a label or icon, and in various states.
+---
