+++ site/docs/components/button/usage.mdx
@@ -1,0 +1,33 @@
+---
+title: Button
+layout: DetailComponent
+sidebar:
+  label: How to use
+data:
+  descriptionRef:
+    $ref: ./#/data/description
+---
+
+### Using the Button component
+
+- Keep the text label short. Ideally one to three words and no more than five. If you’re including an icon in the label, use no more than three words.
+
+- When including icons, use the Salt Figma icon library. If the icon you need isn’t available, follow the steps in our Creating Salt icons guide.
+
+- Consider a Link instead of the Button component if the user is navigating to a new page or window rather than carrying out an action.
+
+### Import
+
+To import Button from the core Salt package, use:
+
+```
+import { Button } from "@salt-ds/core";
+```
+
+### Props
+
+| Name                  | Type      | Description                                                                                          | Default   |
+| --------------------- | --------- | ---------------------------------------------------------------------------------------------------- | --------- |
+| disabled              | `boolean` | If `true``, the button will be disabled.                                                             | -         |
+| focusableWhenDisabled | `boolean` | If `true``, the button will be fosusable when disabled.                                              | -         |
+| variant               | -         | The variant to use. “Primary” is the default value, and the other options are “secondary” and “cta.” | "primary" |
