+++ site/snapshots/latest/salt/components/button/accessibility
@@ -1,0 +1,1 @@
+accessibility.mdx
