+++ site/snapshots/latest/salt/components/button/accessibility.mdx
@@ -1,0 +1,88 @@
+---
+title: Button
+layout: DetailComponent
+sidebar:
+  label: Accessibility
+data:
+  descriptionRef: >-
+    Button is an interactive component that allows the user to execute an
+    action. There are three Button variants—CTA (call to action), primary and
+    secondary—that you can display with or without a label or icon, and in
+    various states.
+lastModified: 1681920447574
+fullPath: /salt/components/button/accessibility.mdx
+route: /salt/components/button/accessibility
+readingTime:
+  text: 1 min read
+  minutes: 0.79
+  time: 47400
+  words: 158
+tableOfContents:
+  - level: 3
+    id: best-practises
+    text: Best practises
+  - level: 3
+    id: keyboard-interactions
+    text: Keyboard interactions
+navigation:
+  prev:
+    title: Button
+    route: /salt/components/button/index
+  next:
+    title: Button
+    route: /salt/components/button/examples
+sidebarData:
+  - id: /salt/components/index
+    fullPath: /salt/components/index.mdx
+    name: Components
+    data:
+      level: 2
+      link: /salt/components/index
+    childNodes:
+      - id: /salt/components/button/index
+        fullPath: /salt/components/button/index.mdx
+        name: Button
+        data:
+          level: 3
+          link: /salt/components/button/index
+        childNodes:
+          - id: /salt/components/button/accessibility
+            fullPath: /salt/components/button/accessibility.mdx
+            name: Accessibility
+            data:
+              level: 3
+              link: /salt/components/button/accessibility
+            childNodes: []
+          - id: /salt/components/button/examples
+            fullPath: /salt/components/button/examples.mdx
+            name: Examples
+            data:
+              level: 3
+              link: /salt/components/button/examples
+            childNodes: []
+          - id: /salt/components/button/usage
+            fullPath: /salt/components/button/usage.mdx
+            name: How to use
+            data:
+              level: 3
+              link: /salt/components/button/usage
+            childNodes: []
+---
+
+### Best practises
+
+#### When a Button has both an icon and a label
+
+- Pass `aria-hidden=”true”` to the icon so that its label is not announced by screen readers.
+
+#### When a Button has an icon only and no label
+
+- Pass an `aria-label` to Button that describes its purpose.
+
+### Keyboard interactions
+
+| Key                                  | Function                                                                                                                                                                                                                                                                |
+| ------------------------------------ | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| <kbd>Tab</kbd>                       | - If focus is above the button, Tab key press moves focus onto the button.<br/>- If the button is disabled and is set to receive focus, receives focus.<br/>- If the button has focus, Tab moves focus out of the button, to next focusable component in the tab order. |
+| <kbd>Tab</kbd> and <kbd>Shift</kbd>  | - If focused, Tab+Shift combo moves focus out of the component to the previous component in the tab order.                                                                                                                                                              |
+| <kbd>Enter</kbd> or <kbd>Space</kbd> | - If the button has focus, Enter or Space key press activates the button.<br/>- If the button is disabled and can receive focus, does nothing.                                                                                                                          |
