+++ site/snapshots/latest/salt/components/button/examples
@@ -1,0 +1,1 @@
+examples.mdx
