+++ site/snapshots/latest/salt/components/button/examples.mdx
@@ -1,19 +1,33 @@
 ---
-title: Components
-layout: DetailContentOnly
-lastModified: 1681471884713
-fullPath: /salt/components/index.mdx
-route: /salt/components/index
+title: Button
+layout: DetailComponent
+sidebar:
+  label: Examples
+data:
+  descriptionRef: >-
+    Button is an interactive component that allows the user to execute an
+    action. There are three Button variants—CTA (call to action), primary and
+    secondary—that you can display with or without a label or icon, and in
+    various states.
+lastModified: 1681920447575
+fullPath: /salt/components/button/examples.mdx
+route: /salt/components/button/examples
 readingTime:
   text: 1 min read
-  minutes: 0.48
-  time: 28800
-  words: 96
-tableOfContents: []
+  minutes: 0.28
+  time: 16800
+  words: 56
+tableOfContents:
+  - level: 3
+    id: cta-call-to-action
+    text: CTA (call to action)
 navigation:
+  prev:
+    title: Button
+    route: /salt/components/button/accessibility
   next:
     title: Button
-    route: /salt/components/button/index
+    route: /salt/components/button/usage
 sidebarData:
   - id: /salt/components/index
     fullPath: /salt/components/index.mdx
@@ -52,15 +66,12 @@ sidebarData:
             childNodes: []
 ---
 
-Our components are built with accessibility in mind. Each one has comprehensive
-supporting documentation with usage guidance, examples and props.
-We’re constantly adding new components so check back if what you need isn’t
-available or [get in touch](../support-and-contributions/index) if you have any feedback or requests.
+### CTA (call to action)
 
-In addition to the public libraries in the [Figma Community](https://www.figma.com/@jpmorgan_salt),
-if you’re a J.P. Morgan employee, you can view the [Salt component files](https://www.figma.com/files/project/72279506)
-in Figma when logged in to the JPMC network.
-They include accessibility and design specifications for each component,
-with designs for light and dark modes in all four densities.
+Use the CTA button for high priority actions.
 
-<ComponentsList />
+#### Guidance
+
+- Choose a CTA button when an action is more urgent than other visible buttons.
+- Use it for important or irreversible actions, such as agreeing to terms and conditions or deleting a profile.
+- Only display one CTA button per screen to emphasize the next action the user should take.
