+++ site/snapshots/latest/salt/components/button/index
@@ -1,0 +1,1 @@
+index.mdx
