+++ site/snapshots/latest/salt/components/button/index.mdx
@@ -1,19 +1,24 @@
 ---
-title: Components
-layout: DetailContentOnly
-lastModified: 1681471884713
-fullPath: /salt/components/index.mdx
-route: /salt/components/index
+title: Button
+layout: DetailComponent
+data:
+  description: >-
+    Button is an interactive component that allows the user to execute an
+    action. There are three Button variants—CTA (call to action), primary and
+    secondary—that you can display with or without a label or icon, and in
+    various states.
+lastModified: 1681920447576
+fullPath: /salt/components/button/index.mdx
+route: /salt/components/button/index
 readingTime:
-  text: 1 min read
-  minutes: 0.48
-  time: 28800
-  words: 96
-tableOfContents: []
+  text: 0 min read
+  minutes: 0
+  time: 0
+  words: 0
 navigation:
   next:
     title: Button
-    route: /salt/components/button/index
+    route: /salt/components/button/accessibility
 sidebarData:
   - id: /salt/components/index
     fullPath: /salt/components/index.mdx
@@ -51,16 +56,3 @@ sidebarData:
               link: /salt/components/button/usage
             childNodes: []
 ---
-
-Our components are built with accessibility in mind. Each one has comprehensive
-supporting documentation with usage guidance, examples and props.
-We’re constantly adding new components so check back if what you need isn’t
-available or [get in touch](../support-and-contributions/index) if you have any feedback or requests.
-
-In addition to the public libraries in the [Figma Community](https://www.figma.com/@jpmorgan_salt),
-if you’re a J.P. Morgan employee, you can view the [Salt component files](https://www.figma.com/files/project/72279506)
-in Figma when logged in to the JPMC network.
-They include accessibility and design specifications for each component,
-with designs for light and dark modes in all four densities.
-
-<ComponentsList />
