+++ site/snapshots/latest/salt/components/button/shared-config.json
@@ -1,0 +1,1 @@
+../shared-config.json
