+++ site/snapshots/latest/salt/components/button/usage.mdx
@@ -1,0 +1,94 @@
+---
+title: Button
+layout: DetailComponent
+sidebar:
+  label: How to use
+data:
+  descriptionRef: >-
+    Button is an interactive component that allows the user to execute an
+    action. There are three Button variants—CTA (call to action), primary and
+    secondary—that you can display with or without a label or icon, and in
+    various states.
+lastModified: 1681920447579
+fullPath: /salt/components/button/usage.mdx
+route: /salt/components/button/usage
+readingTime:
+  text: 1 min read
+  minutes: 0.755
+  time: 45300
+  words: 151
+tableOfContents:
+  - level: 3
+    id: using-the-button-component
+    text: Using the Button component
+  - level: 3
+    id: import
+    text: Import
+  - level: 3
+    id: props
+    text: Props
+navigation:
+  prev:
+    title: Button
+    route: /salt/components/button/examples
+sidebarData:
+  - id: /salt/components/index
+    fullPath: /salt/components/index.mdx
+    name: Components
+    data:
+      level: 2
+      link: /salt/components/index
+    childNodes:
+      - id: /salt/components/button/index
+        fullPath: /salt/components/button/index.mdx
+        name: Button
+        data:
+          level: 3
+          link: /salt/components/button/index
+        childNodes:
+          - id: /salt/components/button/accessibility
+            fullPath: /salt/components/button/accessibility.mdx
+            name: Accessibility
+            data:
+              level: 3
+              link: /salt/components/button/accessibility
+            childNodes: []
+          - id: /salt/components/button/examples
+            fullPath: /salt/components/button/examples.mdx
+            name: Examples
+            data:
+              level: 3
+              link: /salt/components/button/examples
+            childNodes: []
+          - id: /salt/components/button/usage
+            fullPath: /salt/components/button/usage.mdx
+            name: How to use
+            data:
+              level: 3
+              link: /salt/components/button/usage
+            childNodes: []
+---
+
+### Using the Button component
+
+- Keep the text label short. Ideally one to three words and no more than five. If you’re including an icon in the label, use no more than three words.
+
+- When including icons, use the Salt Figma icon library. If the icon you need isn’t available, follow the steps in our Creating Salt icons guide.
+
+- Consider a Link instead of the Button component if the user is navigating to a new page or window rather than carrying out an action.
+
+### Import
+
+To import Button from the core Salt package, use:
+
+```
+import { Button } from "@salt-ds/core";
+```
+
+### Props
+
+| Name                  | Type      | Description                                                                                          | Default   |
+| --------------------- | --------- | ---------------------------------------------------------------------------------------------------- | --------- |
+| disabled              | `boolean` | If `true``, the button will be disabled.                                                             | -         |
+| focusableWhenDisabled | `boolean` | If `true``, the button will be fosusable when disabled.                                              | -         |
+| variant               | -         | The variant to use. “Primary” is the default value, and the other options are “secondary” and “cta.” | "primary" |
