+++ site/snapshots/latest/salt/components/sidebar.json
@@ -5,7 +5,43 @@
       "fullPath": "/salt/components/index.mdx",
       "name": "Components",
       "data": { "level": 2, "link": "/salt/components/index" },
-      "childNodes": []
+      "childNodes": [
+        {
+          "id": "/salt/components/button/index",
+          "fullPath": "/salt/components/button/index.mdx",
+          "name": "Button",
+          "data": { "level": 3, "link": "/salt/components/button/index" },
+          "childNodes": [
+            {
+              "id": "/salt/components/button/accessibility",
+              "fullPath": "/salt/components/button/accessibility.mdx",
+              "name": "Accessibility",
+              "data": {
+                "level": 3,
+                "link": "/salt/components/button/accessibility"
+              },
+              "childNodes": []
+            },
+            {
+              "id": "/salt/components/button/examples",
+              "fullPath": "/salt/components/button/examples.mdx",
+              "name": "Examples",
+              "data": {
+                "level": 3,
+                "link": "/salt/components/button/examples"
+              },
+              "childNodes": []
+            },
+            {
+              "id": "/salt/components/button/usage",
+              "fullPath": "/salt/components/button/usage.mdx",
+              "name": "How to use",
+              "data": { "level": 3, "link": "/salt/components/button/usage" },
+              "childNodes": []
+            }
+          ]
+        }
+      ]
     }
   ]
 }
