+++ site/snapshots/latest/salt/getting-started/designing.mdx
@@ -3,7 +3,7 @@ title: Designing with Salt
 layout: DetailTechnical
 sidebar:
   label: Designing
-lastModified: 1680709263986
+lastModified: 1681471884714
 fullPath: /salt/getting-started/designing.mdx
 route: /salt/getting-started/designing
 readingTime:
