+++ site/snapshots/latest/salt/getting-started/developing.mdx
@@ -3,7 +3,7 @@ title: Developing with Salt
 layout: DetailTechnical
 sidebar:
   label: Developing
-lastModified: 1680709263986
+lastModified: 1681471884714
 fullPath: /salt/getting-started/developing.mdx
 route: /salt/getting-started/developing
 readingTime:
