+++ site/snapshots/latest/salt/getting-started/index.mdx
@@ -3,7 +3,7 @@ title: Getting started
 layout: DetailTechnical
 sidebar:
   label: Overview
-lastModified: 1680709263987
+lastModified: 1681471884715
 fullPath: /salt/getting-started/index.mdx
 route: /salt/getting-started/index
 readingTime:
