+++ site/snapshots/latest/salt/getting-started/roadmap.mdx
@@ -1,7 +1,7 @@
 ---
 title: Roadmap
 layout: DetailTechnical
-lastModified: 1680709263987
+lastModified: 1681471884715
 fullPath: /salt/getting-started/roadmap.mdx
 route: /salt/getting-started/roadmap
 readingTime:
