+++ site/snapshots/latest/salt/index.mdx
@@ -25,7 +25,7 @@ frameOverrides: &ref_0
         to: "https://www.jpmorgan.com/privacy"
       - label: Contact us
         to: /salt/support-and-contributions/index
-lastModified: 1680709263731
+lastModified: 1681471885098
 fullPath: /salt/index.mdx
 route: /salt/index
 sharedConfig: *ref_0
