+++ site/snapshots/latest/salt/support-and-contributions/index.mdx
@@ -1,7 +1,7 @@
 ---
 title: Support and contributions
 layout: DetailContentOnly
-lastModified: 1680709263988
+lastModified: 1681471884716
 fullPath: /salt/support-and-contributions/index.mdx
 route: /salt/support-and-contributions/index
 readingTime:
