+++ site/snapshots/latest/search-data.json
@@ -412,5 +412,83 @@
       "If your team needs a feature or functionality that’s not currently provided, we may be able to assist.",
       "Email us if you have a complex product or design-related query and we’ll do our best to help."
     ]
+  },
+  {
+    "title": "Button",
+    "route": "/salt/components/button/accessibility",
+    "content": [
+      "Best practises",
+      "When a Button has both an icon and a label",
+      "Pass ",
+      " to the icon so that its label is not announced by screen readers.",
+      "When a Button has an icon only and no label",
+      "Pass an ",
+      " to Button that describes its purpose.",
+      "Keyboard interactions",
+      "| Key                                  | Function                                                                                                                                                                                               ",
+      "| ------------------------------------ | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------",
+      "| ",
+      "Tab",
+      "                       | - If focus is above the button, Tab key press moves focus onto the button.",
+      "- If the button is disabled and is set to receive focus, receives focus.",
+      "- If the button has focus, Tab moves focus out of the button, to next focusable component in the tab order. ",
+      "|\n",
+      "| ",
+      "Tab",
+      " and ",
+      "Shift",
+      "  | - If focused, Tab+Shift combo moves focus out of the component to the previous component in the tab order.                                                                                                                                  ",
+      "|\n",
+      "| ",
+      "Enter",
+      " or ",
+      "Space",
+      " | - If the button has focus, Enter or Space key press activates the button.",
+      "- If the button is disabled and can receive focus, does nothing.                                                                                                                          ",
+      "|"
+    ]
+  },
+  {
+    "title": "Button",
+    "route": "/salt/components/button/examples",
+    "content": [
+      "CTA (call to action)",
+      "Use the CTA button for high priority actions.",
+      "Guidance",
+      "Choose a CTA button when an action is more urgent than other visible buttons.",
+      "Use it for important or irreversible actions, such as agreeing to terms and conditions or deleting a profile.",
+      "Only display one CTA button per screen to emphasize the next action the user should take."
+    ]
+  },
+  {
+    "title": "Button",
+    "route": "/salt/components/button/index",
+    "content": []
+  },
+  {
+    "title": "Button",
+    "route": "/salt/components/button/usage",
+    "content": [
+      "Using the Button component",
+      "Keep the text label short. ",
+      "Ideally one to three words and no more than five. ",
+      "If you’re including an icon in the label, use no more than three words.",
+      "When including icons, use the Salt Figma icon library. ",
+      "If the icon you need isn’t available, follow the steps in our Creating Salt icons guide.",
+      "Consider a Link instead of the Button component if the user is navigating to a new page or window rather than carrying out an action.",
+      "Import",
+      "To import Button from the core Salt package, use:",
+      "import { Button } from \"@salt-ds/core\";",
+      "Props",
+      "| Name                  | Type      | Description                                                                                          | Default   |\n",
+      "| --------------------- | --------- | ---------------------------------------------------------------------------------------------------- | --------- |\n",
+      "| disabled              | ",
+      " | If ",
+      "boolean",
+      "true``, the button will be fosusable when disabled.                                              ",
+      "| -         |\n",
+      "| variant               | -         | The variant to use. ",
+      "“Primary” is the default value, and the other options are “secondary” and “cta.” | \"primary\" |"
+    ]
   }
 ]
