+++ site/src/layouts/DetailBase/DetailBase.module.css
