+++ site/src/layouts/DetailBase/index.ts
@@ -1,0 +1,1 @@
+export * from "./DetailBase";
