+++ site/src/layouts/types/index.ts
@@ -7,6 +7,7 @@ import { FooterProps } from "@jpmorganchase/mosaic-site-components";
 
 export type LayoutProps = {
   children?: ReactNode;
+  sidebar?: ReactElement;
   className?: string;
   ToCProps?: {
     items: any;
